Create handlers module and clean up
Refactored the handlers logic from bot.js to handlers/pingHandler and
handlers/leaguePatchNotesHandler. It should be cleaner if the handler
only exposes the command dictionary that the bot uses to run the
associated handler.

Also, the package.json file just had a bad name field. Having a space or
starting with a capital letter makes npm install fail.